requirements.txt
amqp==2.6.1
asgiref==3.2.10
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.14.34
botocore==1.17.34
bs4==0.0.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.1
docutils==0.15.2
flower==0.9.5
humanize==2.5.0
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
kombu==4.6.11
Markdown==3.2.2
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
prometheus-client==0.8.0
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
